SONAR-27806 Refactor tests
To increase readability and make it easier to add new tests
